requirements.txt
# LangChain核心库
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10

# 嵌入模型
openai>=1.0.0
tiktoken>=0.5.0

# 向量数据库
chromadb>=0.4.0
faiss-cpu>=1.7.0

# 文档处理
pypdf>=3.0.0
python-docx>=0.8.11
unstructured>=0.10.0

# 实用工具
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# 测试框架
pytest>=7.0.0
pytest-mock>=3.10.0

# 环境变量管理
python-dotenv>=1.0.0

# 性能监控
psutil>=5.9.0